requirements.txt
langchain
pypdf2
python-dotenv
openai
fastapi  
uvicorn
sentence_transformers>=3.0.0
transformers<=4.50.0
pdfplumber
PyMuPDF
diskcache
textstat
ffmpeg-python
python-multipart
moviepy 
huggingface_hub
pybase64
reportlab